refactor(progress-bar-web): add percentage label option
